Add "subproviders" and "search platforms"
A "subprovider" is a service that will be supported when the user send a link a associated to them, a "search platform" is a service that can be used by the bot to make a search. For example, this commit give users the ability to send an Apple Music link that will be used to search and download the music through YouTube.